requirements.txt
python-dotenv
supabase
aiohttp
time
csv
shutil
asyncio
aiohttp